+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     * Fixed modular form submits without defined `action: /path` inside the form
     * Fixed form processing in nested modular pages
     * Fixed file field saving with nested name
+    * Fixed file saving if destination folder does not exist
 
 # v3.0.0-beta.5
 ## 01/25/2019